+++ internal/utils/tmplkit/internal.go
@@ -1,4 +1,4 @@
-package templatefunc
+package tmplkit
 
 import (
 	"strings"
